+++ src/app/features/create-account/workplace/is-this-your-workplace/is-this-your-workplace.component.ts
@@ -1,108 +1,26 @@
-import { AfterViewInit, Component, ElementRef, OnInit, ViewChild } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { Component } from '@angular/core';
+import { FormBuilder } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
-import { ErrorDefinition, ErrorDetails } from '@core/model/errorSummary.model';
-import { Establishment } from '@core/model/establishment.model';
-import { LocationAddress } from '@core/model/location.model';
 import { BackService } from '@core/services/back.service';
 import { ErrorSummaryService } from '@core/services/error-summary.service';
 import { EstablishmentService } from '@core/services/establishment.service';
 import { RegistrationService } from '@core/services/registration.service';
+import { IsThisYourWorkplaceDirective } from '@shared/directives/create-workplace/is-this-your-workplace/is-this-your-workplace.directive';
 
 @Component({
   selector: 'app-is-this-your-workplace',
   templateUrl: './is-this-your-workplace.component.html',
 })
-export class IsThisYourWorkplaceComponent implements OnInit, AfterViewInit {
-  @ViewChild('formEl') formEl: ElementRef;
-  public form: FormGroup;
-  public formErrorsMap: Array<ErrorDetails>;
-  protected serverErrorsMap: Array<ErrorDefinition>;
-  public submitted = false;
-  private flow: string;
-  public locationData: LocationAddress;
-  public serverError: string;
-  public workplace: Establishment;
-  public isParent: boolean;
-  public searchMethod: string;
-
+export class IsThisYourWorkplaceComponent extends IsThisYourWorkplaceDirective {
   constructor(
-    private errorSummaryService: ErrorSummaryService,
-    private establishmentService: EstablishmentService,
+    protected errorSummaryService: ErrorSummaryService,
+    protected establishmentService: EstablishmentService,
     public backService: BackService,
-    private route: ActivatedRoute,
-    private router: Router,
-    private registrationService: RegistrationService,
-    private formBuilder: FormBuilder,
-  ) {}
-
-  ngOnInit(): void {
-    this.flow = this.route.snapshot.parent.url[0].path;
-    this.setupForm();
-    this.setBackLink();
-    this.locationData = this.registrationService.locationAddresses$.value[0];
-    this.searchMethod = this.registrationService.searchMethod$.value;
-    this.workplace = this.establishmentService.primaryWorkplace;
-    this.workplace?.isParent ? (this.isParent = true) : (this.isParent = false);
-    this.setupFormErrorsMap();
-  }
-
-  public ngAfterViewInit(): void {
-    this.errorSummaryService.formEl$.next(this.formEl);
-  }
-
-  private setupForm(): void {
-    this.form = this.formBuilder.group({
-      yourWorkplace: [
-        null,
-        {
-          validators: [Validators.required],
-          updateOn: 'submit',
-        },
-      ],
-    });
-  }
-
-  private setupFormErrorsMap(): void {
-    let errorMessage;
-    this.isParent
-      ? (errorMessage = 'Select yes if this is the workplace you want to add')
-      : (errorMessage = 'Select yes if this is your workplace');
-
-    this.formErrorsMap = [
-      {
-        item: 'yourWorkplace',
-        type: [
-          {
-            name: 'required',
-            message: errorMessage,
-          },
-        ],
-      },
-    ];
-  }
-
-  public getErrorMessage(item: string): string {
-    const errorType = Object.keys(this.form.get(item).errors)[0];
-    return this.errorSummaryService.getFormErrorMessage(item, errorType, this.formErrorsMap);
-  }
-
-  public setBackLink(): void {
-    this.backService.setBackLink({ url: [`/${this.flow}`, 'find-workplace'] });
-  }
-
-  onSubmit(): void {
-    const yourWorkplace = this.form.get('yourWorkplace');
-    this.submitted = true;
-
-    if (this.form.valid) {
-      if (yourWorkplace.value === 'yes') {
-        this.router.navigate([this.flow, 'new-select-main-service']);
-      } else {
-        this.router.navigate([this.flow, 'find-workplace']);
-      }
-    } else {
-      this.errorSummaryService.scrollToErrorSummary();
-    }
+    protected route: ActivatedRoute,
+    protected router: Router,
+    protected registrationService: RegistrationService,
+    protected formBuilder: FormBuilder,
+  ) {
+    super(errorSummaryService, establishmentService, backService, route, router, registrationService, formBuilder);
   }
 }
